Updated households tests and consolidated them into the FixedRecordTests.java to improve testing time.
